+++ azure_sdk_core/src/lib.rs
@@ -721,7 +721,7 @@ pub trait ContentMD5Option<'a> {
 
 #[inline]
 #[must_use]
-pub fn add_content_md5_header<'a>(content_md5: &'a [u8], builder: Builder) -> Builder {
+pub fn add_content_md5_header(content_md5: &[u8], builder: Builder) -> Builder {
     let s = encode(content_md5);
     builder.header(CONTENT_MD5, &s as &str)
 }
