+++ azure_sdk_cosmos/src/responses/create_slug_attachment_response.rs
@@ -48,7 +48,7 @@ impl std::convert::TryFrom<(&HeaderMap, &[u8])> for CreateSlugAttachmentResponse
             max_media_storage_usage_mb: max_media_storage_usage_mb_from_headers(headers)?,
             media_storage_usage_mb: media_storage_usage_mb_from_headers(headers)?,
             last_change: last_state_change_from_headers(headers)?,
-            etag: etag_from_headers(headers)?.to_owned(),
+            etag: etag_from_headers(headers)?,
             resource_quota: resource_quota_from_headers(headers)?,
             resource_usage: resource_usage_from_headers(headers)?,
             lsn: lsn_from_headers(headers)?,
