+++ azure_sdk_cosmos/src/responses/replace_reference_attachment_response.rs
@@ -45,7 +45,7 @@ impl std::convert::TryFrom<(&HeaderMap, &[u8])> for ReplaceReferenceAttachmentRe
         Ok(Self {
             attachment,
             last_change: last_state_change_from_headers(headers)?,
-            etag: etag_from_headers(headers)?.to_owned(),
+            etag: etag_from_headers(headers)?,
             resource_quota: resource_quota_from_headers(headers)?,
             resource_usage: resource_usage_from_headers(headers)?,
             lsn: lsn_from_headers(headers)?,
